+++ ott/src/components/common/ImagePreview.jsx
